backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
pydicom>=2.4.0
Pillow>=10.0.0
numpy>=1.24.0
pydantic>=2.0.0
